modify some files for large-scale models

+++ src/ComplexSparseSquareMatrix.cpp
@@ -333,11 +333,11 @@ void ComplexSparseSquareMatrix::factorizationPhaseMatrixSolver(){
 }
 
 //Solve phase of matrix solver with a specified number of right-hand side
-void ComplexSparseSquareMatrix::solvePhaseMatrixSolver( std::complex<double>* solution, const int iRhsStart ,const int nRhs ){
+void ComplexSparseSquareMatrix::solvePhaseMatrixSolver( std::complex<double>* solution, const long long iRhsStart ,const int nRhs ){
 
 	assert( m_hasConvertedToCRSFormat );
 
-	const long long index = static_cast<long long>(m_numRows) * static_cast<long long>(iRhsStart);
+	const long long index = static_cast<long long>(m_numRows) * iRhsStart;
 	m_pardisoSolver.solve( m_rowIndex, m_columns, m_values, nRhs, &m_rightHandSideVector[index], solution );
 }
